+++ sources/Editor/GuiComponents/Layout.cpp
@@ -34,7 +34,7 @@ namespace Core::Gui
         {
             if (child->isEnabled())
             {
-                if ((child->unsafeCastTo<Widget>()->getFlex() & f) != 0)
+                if ((static_cast<int>(child->unsafeCastTo<Widget>()->getFlex()) & static_cast<int>(f)) != 0)
                 {
                     return true;
                 }
